fix: sqlite test
