+++ libs/jwst-storage/src/storage/blobs.rs
@@ -29,32 +29,28 @@ impl BlobAutoStorage {
         Self::init_with_pool(pool, get_bucket(is_sqlite)).await
     }
 
-    pub async fn all(&self, table: &str) -> Result<Vec<BlobModel>, DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn all(&self, table: &str) -> Result<Vec<BlobModel>, DbErr> {
         Blobs::find()
             .filter(BlobColumn::Workspace.eq(table))
             .all(&self.pool)
             .await
     }
 
-    pub async fn count(&self, table: &str) -> Result<u64, DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn count(&self, table: &str) -> Result<u64, DbErr> {
         Blobs::find()
             .filter(BlobColumn::Workspace.eq(table))
             .count(&self.pool)
             .await
     }
 
-    pub async fn exists(&self, table: &str, hash: &str) -> Result<bool, DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn exists(&self, table: &str, hash: &str) -> Result<bool, DbErr> {
         Blobs::find_by_id((table.into(), hash.into()))
             .count(&self.pool)
             .await
             .map(|c| c > 0)
     }
 
-    pub async fn metadata(&self, table: &str, hash: &str) -> Result<BlobMetadata, DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn metadata(&self, table: &str, hash: &str) -> Result<BlobMetadata, DbErr> {
         #[derive(FromQueryResult)]
         struct Metadata {
             size: i64,
@@ -76,8 +72,7 @@ impl BlobAutoStorage {
         })
     }
 
-    pub async fn insert(&self, table: &str, hash: &str, blob: &[u8]) -> Result<(), DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn insert(&self, table: &str, hash: &str, blob: &[u8]) -> Result<(), DbErr> {
         if !self.exists(table, hash).await? {
             Blobs::insert(BlobActiveModel {
                 workspace: Set(table.into()),
@@ -93,24 +88,21 @@ impl BlobAutoStorage {
         Ok(())
     }
 
-    pub async fn get(&self, table: &str, hash: &str) -> Result<BlobModel, DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn get(&self, table: &str, hash: &str) -> Result<BlobModel, DbErr> {
         Blobs::find_by_id((table.into(), hash.into()))
             .one(&self.pool)
             .await
             .and_then(|r| r.ok_or(DbErr::Query(RuntimeErr::Internal("blob not exists".into()))))
     }
 
-    pub async fn delete(&self, table: &str, hash: &str) -> Result<bool, DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn delete(&self, table: &str, hash: &str) -> Result<bool, DbErr> {
         Blobs::delete_by_id((table.into(), hash.into()))
             .exec(&self.pool)
             .await
             .map(|r| r.rows_affected == 1)
     }
 
-    pub async fn drop(&self, table: &str) -> Result<(), DbErr> {
-        let _lock = self.bucket.get_lock().await;
+    async fn drop(&self, table: &str) -> Result<(), DbErr> {
         Blobs::delete_many()
             .filter(BlobColumn::Workspace.eq(table))
             .exec(&self.pool)
@@ -125,6 +117,7 @@ impl BlobStorage for BlobAutoStorage {
     type Read = ReaderStream<Cursor<Vec<u8>>>;
 
     async fn get_blob(&self, workspace: Option<String>, id: String) -> JwstResult<Self::Read> {
+        let _lock = self.bucket.get_lock().await;
         let workspace = workspace.unwrap_or("__default__".into());
         if let Ok(blob) = self.get(&workspace, &id).await {
             return Ok(ReaderStream::new(Cursor::new(blob.blob)));
@@ -137,6 +130,7 @@ impl BlobStorage for BlobAutoStorage {
         workspace: Option<String>,
         id: String,
     ) -> JwstResult<BlobMetadata> {
+        let _lock = self.bucket.get_lock().await;
         let workspace = workspace.unwrap_or("__default__".into());
         if let Ok(metadata) = self.metadata(&workspace, &id).await {
             Ok(metadata)
@@ -149,6 +143,7 @@ impl BlobStorage for BlobAutoStorage {
         workspace: Option<String>,
         stream: impl Stream<Item = Bytes> + Send,
     ) -> JwstResult<String> {
+        let _lock = self.bucket.get_lock().await;
         let workspace = workspace.unwrap_or("__default__".into());
 
         let (hash, blob) = get_hash(stream).await;
@@ -160,6 +155,7 @@ impl BlobStorage for BlobAutoStorage {
         }
     }
     async fn delete_blob(&self, workspace_id: Option<String>, id: String) -> JwstResult<()> {
+        let _lock = self.bucket.get_lock().await;
         let workspace_id = workspace_id.unwrap_or("__default__".into());
         if let Ok(_success) = self.delete(&workspace_id, &id).await {
             Ok(())
@@ -168,6 +164,7 @@ impl BlobStorage for BlobAutoStorage {
         }
     }
     async fn delete_workspace(&self, workspace_id: String) -> JwstResult<()> {
+        let _lock = self.bucket.get_lock().await;
         if self.drop(&workspace_id).await.is_ok() {
             Ok(())
         } else {
